refactor: 날짜 입력, 날짜시간 입력 부분 메서드 분리
inputDate() inputDateTime()
