+++ src/main/java/kr/co/fastcampus/travel/view/InputView.java
@@ -3,6 +3,10 @@ package kr.co.fastcampus.travel.view;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 import java.util.function.Predicate;
 import kr.co.fastcampus.travel.common.exception.UnknownException;
 
@@ -11,6 +15,35 @@ public class InputView {
 
     public InputView() {
         this.br = new BufferedReader(new InputStreamReader(System.in));
+    }
+
+    public LocalDateTime inputDateTime() {
+        System.out.println("날짜와 시간 입력 (Enter로 생략 가능, YYYY-MM-DD HH:MM 형식으로 입력):");
+        String dateTimeStr = inputString(
+            str -> str.isEmpty() || isValidDateTime(str),
+            "0000-00-00 00:00 날짜와 시간 형식에 맞지 않습니다. 다시 입력해주세요"
+        );
+        return dateTimeStr.isEmpty()
+            ? null
+            : LocalDateTime.parse(dateTimeStr, DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
+    }
+
+    public LocalDate inputDate() {
+        System.out.println("날짜 입력 (YYYY-MM-DD 형식으로 입력):");
+        String dateStr = inputNotEmptyString(
+            this::isValidDate,
+            "0000-00-00 날짜 형식에 맞지 않습니다. 다시 입력해주세요"
+        );
+        return LocalDate.parse(dateStr, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+    }
+
+    public Menu inputMenu() {
+        try {
+            int menuNumber = inputNumber("잘못된 메뉴 번호입니다. 다시 입력해주세요;");
+            return Menu.fromNumber(menuNumber);
+        } catch (IllegalArgumentException e) {
+            return inputMenu();
+        }
     }
 
     public String inputString(Predicate<String> isValid, String errorMessage) {
@@ -38,15 +71,6 @@ public class InputView {
         return inputNotEmptyString(str -> true, "");
     }
 
-    public Menu inputMenu() {
-        try {
-            int menuNumber = inputNumber("잘못된 메뉴 번호입니다. 다시 입력해주세요;");
-            return Menu.fromNumber(menuNumber);
-        } catch (IllegalArgumentException e) {
-            return inputMenu();
-        }
-    }
-
     public int inputNumber(String errorMessage) {
         while (true) {
             try {
@@ -54,6 +78,24 @@ public class InputView {
             } catch (IllegalArgumentException e) {
                 System.out.println(errorMessage);
             }
+        }
+    }
+
+    private boolean isValidDateTime(String dateTimeStr) {
+        try {
+            LocalDateTime.parse(dateTimeStr, DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
+            return true;
+        } catch (DateTimeParseException e) {
+            return false;
+        }
+    }
+
+    private boolean isValidDate(String dateStr) {
+        try {
+            LocalDate.parse(dateStr, DateTimeFormatter.ofPattern("yyyy-MM-dd"));
+            return true;
+        } catch (DateTimeParseException e) {
+            return false;
         }
     }
